Dragging in the range chart now updates annotation
We've created a delegate protocol which contains a method which is fired
when the annotation is dragged. This is used to update the location of
the main chart.

The gesture recogniser is added to the glView subview because
annotations appear to be too small to register panning events (they are
1x1 px, with an affine transform). I suspect this will require fixing in
a later commit.

Don't currently have handles, or momentum.