Added simple handles on the range selector
These are really simple SChartAnnotation subclasses which add some lines
and guff. We add gesture recognisers to them to allow changing of each
end of the range selection independently.

We've also taken this opportunity to refactor the other range selector
gesture recognition, by creating another SChartAnnotation subclass which
dynamically changes it's size on transform updates rather than applying
the transform. This means that the annotation will have a size which
will allow pan gestures to be triggered within it.

The gesture code in the annotation manager needs refactoring - at the
moment the method handleHandlePan: is not particularly pleasantly named.