First pass at adding momentum to range selector
We create a class which is responsible for generic animation with
normalised distances. This is invoked when the pan gesture recogniser
has ended.

There is some space for refactoring since there is copy/pasted code in
the range manager. The endPosition calculation should be checked on a
device as well - this is very much down to the look and feel.