Adds a simple value line annotation
This follows the value with the highest yValue currently visible on the
chart.

We create a ValueAnnotationManager which controls the display and
positioning of the annotation. It needs to be informed when the
displayed range on the xAxis changes. Currently only horizontal lines
are shown, since we're dealing with timeseries. It would be trivial to
change to allow vertical lines too, but would require a series
monotonically increasing in y rather than x.

As part of this we've created a simple binary search for NSArray which
doesn't require the element you're searching for to exist in the array.
There are 2 methods, and some pretty basic unit tests.