Adds text to the value annotation
We've created a subclass of SChartAnnotation to allow anchoring at a
point other than the centre (here we anchor to the bottom right).

We've also updated how the location of the value annotation is
established by removing reliance on internal ShinobiChart methods. This
involves creation of a new protocol which allows a datasource to return
a y-value for a given x-value. We implement this and update the
RangeSelector appropriately.